Don't use memory that is both writable and executable
The main reason to do this is that it is more secure, as memory
that is both writable and executable is a security risk. It is
also portable to systems that do not allow memory that is both
writable and executable (OpenBSD by default, NetBSD and Linux
with security patches).

Initialize memory as read-execute.  While writing, mprotect it to
read-write, then back to read-execute afterward.

There may be other places that allow memory writing besides write_to
and this would break those cases.  I didn't audit the code to look
for additional spots where you would need to call allow_writes.

Not sure whether this works.  The tests don't run on OpenBSD and
the example segfaults with or without this change.